+++ packages/wallets/src/evm/connectors/magic-eden/getInjectedMagicEdenProvider.ts
@@ -1,0 +1,26 @@
+import { Ethereum } from "../injected/types";
+
+declare global {
+  interface Window {
+    magicEden?: {
+      ethereum?: Ethereum;
+    };
+  }
+}
+
+/**
+ * @internal
+ */
+export function getInjectedMagicEdenProvider(): Ethereum | undefined {
+  if (typeof window === "undefined") {
+    return;
+  }
+
+  if (
+    globalThis.window &&
+    globalThis.window.magicEden &&
+    globalThis.window.magicEden.ethereum
+  ) {
+    return globalThis.window.magicEden.ethereum;
+  }
+}